apps/agents/pyproject.toml
[project]
name = "voicerag-agents"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents>=1.2.0",
    "livekit-plugins-deepgram>=1.0.0",
    "livekit-plugins-openai>=1.0.0",
    "livekit-plugins-silero>=1.0.0",
    "edge-tts>=7.0.0",
    "asyncpg>=0.30.0",
    "httpx>=0.27.2",
    "python-dotenv>=1.0.1",
    "openai>=1.50.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "livekit-agents[testing]>=1.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
